Moves temporary_fstream to ext::fstream & exposes get_matching_set
